Fixed by remarks.
1) <> GridRect -> GridRectangle.
2) -> GridREctangle to Parallelepiped moved back.
3) Refactoring.
