+++ src/com/company/gamecontent/Unit.java
@@ -10,7 +10,7 @@ import com.company.gametools.MathTools;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import com.company.gamegeom.GridRect;
+import com.company.gamegeom.Parallelepiped.GridRectangle;
 
 import static com.company.gamecontent.Restrictions.BLOCK_SIZE;
 import static com.company.gamecontent.Restrictions.INTERSECTION_STRATEGY_SEVERITY;
@@ -290,7 +290,7 @@ public class Unit extends GameObject implements Shootable {
         // So we actually "crop" the rectangle detectionAreaRect with the map rectangle
         detectionAreaRect = GameMap.getInstance().crop(detectionAreaRect);
 
-        GridRect gridRect = new GridRect(detectionAreaRect);
+        GridRectangle gridRect = new GridRectangle(detectionAreaRect);
 //        LOG.debug("Player " + this.getPlayerId() + ": left=" + left + ", right=" + right + ", top=" + top + ", bottom=" + bottom);
 
         // TODO Use Collections
